fix: add indexes to campaigns and foreign key constraint

+++ backend/config/initializers/active_storage.rb
@@ -1,16 +1,19 @@
 # config/initializers/active_storage.rb
 Rails.application.config.to_prepare do
-  # Make purge job more resilient
   ActiveStorage::PurgeJob.class_eval do
     discard_on ActiveRecord::RecordNotFound
     discard_on Aws::S3::Errors::NoSuchKey
     
-    retry_on Aws::S3::Errors::ServiceError, wait: 5.seconds, attempts: 3
+    retry_on Aws::S3::Errors::ServiceError, wait: :exponentially_longer, attempts: 3
+    retry_on ActiveStorage::FileNotFoundError, wait: :exponentially_longer, attempts: 3
     
     def perform(blob)
-      blob.purge if blob.service.exist?(blob.key)
-    rescue Aws::S3::Errors::NoSuchKey
-      blob.destroy
+      if blob.service.exist?(blob.key)
+        blob.purge
+      else
+        # If file doesn't exist, just destroy the record
+        blob.destroy
+      end
     end
   end
 end
